毕设代码/requirements.txt
opencv-python==4.5.3.56
requests==2.18.4
torch==1.8.1
torchvision==0.9.1
numpy==1.19.5
scipy==1.5.4
imageio==2.9.0
Pillow==8.3.1
tensorflow-gpu==1.15.0
scikit-learn==1.0
scikit-image==0.18.3